Added SourceSpan struct.
SourceSpan is used to store information about a span of source
code in a uniform way.  Tokens now use SourceSpans to store
information about where in the source the token came from.
SourceSpans will also be used throughout the parsing process
to maintain information about where in the source a part of
the parse tree came from.