Changed Vec<Token> to VecDeque<Token>, so we have pop_front.
This is came about while attempting to write more of the parser,
so there's some other random stuff in this commit.  Basically,
it's a lot easier to write the parser with a pop_front() method
on the list of tokens so that the various parsing functions can
consume the tokens as they go, instead of having to manually
keep track of where we are in the token list.